26/05: fix acc co khu, phong co thiet bi bao khong duoc xoa va khong the xoa acc admin

+++ src/java/controller/deletePhong.java
@@ -1,6 +1,7 @@
 package controller;
 
 import dal.PhongDAO;
+import jakarta.servlet.RequestDispatcher;
 import java.io.IOException;
 import java.util.List;
 import jakarta.servlet.ServletException;
@@ -9,6 +10,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
+import java.util.ArrayList;
 import model.Accounts;
 import model.Phong;
 
@@ -17,35 +19,29 @@ public class deletePhong extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-          HttpSession session = request.getSession();
+        HttpSession session = request.getSession();
         Accounts a = (Accounts) session.getAttribute("acc");
- 
-         if (a == null) {
-            // Redirect to login page or show error message if account is not logged in
+        if (a == null) {
             response.sendRedirect("login.jsp");
         } else {
-        // Lấy số phòng từ request
-        int phongID = Integer.parseInt(request.getParameter("phongID"));
-        
-        // Kiểm tra phòng còn trống hay không
-        PhongDAO phongDAO = new PhongDAO();
-        List<Phong> phongs = phongDAO.getPhongByPhongID(phongID);
-        if (phongs.size() > 0) {
-            Phong phong = phongs.get(0);
-            if (phong.getPhongConTrong() == 1) {
-                // Nếu phòng trống, thực hiện xóa
-                phongDAO.deletePhongByPhongID(phongID);
-                
-                // Đặt thuộc tính để thông báo xóa thành công trên trang Thông Tin Phòng
-                request.setAttribute("deleted", true);
+            int phongID = Integer.parseInt(request.getParameter("phongID"));
+            PhongDAO phongDAO = new PhongDAO();
+            if (phongDAO.checkPhongIDcoThietBi(phongID)) {
+                List<Phong> phongs = phongDAO.getPhongByPhongID(phongID);
+                request.setAttribute("error", "Phòng đang có thiết bị không thể xóa phòng!");
+                request.setAttribute("danhSachPhong", phongs);
+                request.getRequestDispatcher("HienThiThongTinPhong.jsp").forward(request, response);
+            } else {
+                List<Phong> phongs = phongDAO.getPhongByPhongID(phongID);
+                if (phongs.size() > 0) {
+                    Phong phong = phongs.get(0);
+                    if (phong.getPhongConTrong() == 1) {
+                        phongDAO.deletePhongByPhongID(phongID);
+                        request.setAttribute("deleted", true);
+                    }
+                }
             }
         }
-        
-        // Sau khi xóa, chuyển hướng lại trang danh sách phòng
         response.sendRedirect(request.getContextPath() + "/listphong");
     }
-    }
 }
-
-
-
